Too late kadomatsu
A story that the United States says one country pacifism as nationalism. But what is going to be the trade in G8 if it isolates with that ?? You can not make anything other than war without profit from external interest. In other words, how much American citizen understands to take tax from common people ??